separate example camera control
